requirements.txt
torch
openai-whisper
tqdm
ffmpeg-python